+++ library/src/main/java/me/denley/preferenceinjector/internal/PrefType.java
@@ -9,8 +9,8 @@ public enum PrefType {
     FLOAT("float", "getFloat", "0"),
     INTEGER("int", "getInt", "0"),
     LONG("long", "getLong", "0"),
-    STRING("String", "getString", "null"),
-    STRING_SET("Set<String>", "getStringSet", "null");
+    STRING("java.lang.String", "getString", "null"),
+    STRING_SET("java.util.Set<java.lang.String>", "getStringSet", "null");
 
     private String fieldTypeDef;
     private String methodName;