+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+### Fixed
+- `String` and `Set<String>` are now properly recognized as valid types
+
 ## 2.0.0 - 2015-02-19
 ### Changed
 - Removed `initialize` flag from `@OnPreferenceChange` annotation, `@InjectPreference` can now be applied to methods instead